conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
anyio=4.4.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-lru=2.0.4=pypi_0
attrs=24.1.0=pypi_0
babel=2.15.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.4=hbcca054_0
certifi=2024.7.4=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py311h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
fastjsonschema=2.20.0=pypi_0
fonttools=4.53.1=pypi_0
fqdn=1.5.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=8.2.0=pyha770c72_0
importlib_metadata=8.2.0=hd8ed1ab_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.26.0=pyh707e725_0
ipywidgets=8.1.3=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
json5=0.9.25=pypi_0
jsonpickle=3.2.2=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py311h38be061_0
jupyterlab=4.2.4=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=14.1.0=h77fa898_0
libgomp=14.1.0=h77fa898_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mistune=3.0.2=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=pypi_0
notebook=7.2.1=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=pypi_0
openssl=3.3.1=h4bc722e_2
overrides=7.7.0=pypi_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=pypi_0
pip=24.0=py311h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
plotly=5.23.0=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=6.0.0=py311h331c9d8_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pycparser=2.22=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pypi_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-graphviz=0.20.3=pypi_0
python-json-logger=2.0.7=pypi_0
python_abi=3.11=2_cp311
pytz=2024.1=pypi_0
pyvis=0.3.2=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py311h6a678d5_0
qtconsole=5.5.2=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.19.1=pypi_0
scipy=1.14.0=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
setuptools=69.5.1=py311h06a4308_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
tabulate=0.9.0=pypi_0
tenacity=9.0.0=pypi_0
terminado=0.18.1=pypi_0
tikzplotlib=0.10.1=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h39e8969_0
tornado=6.4.1=py311h331c9d8_0
traitlets=5.14.3=pyhd8ed1ab_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.43.0=py311h06a4308_0
widgetsnbextension=4.0.11=pypi_0
xz=5.4.6=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
